use global hook
