+++ src/app/layout.tsx
@@ -1,9 +1,8 @@
 "use client";
 
-import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 import { App, ConfigProvider, Layout, theme } from "antd"
-import React, { PropsWithChildren, useState } from "react";
+import React, { PropsWithChildren, useEffect, useState } from "react";
 import StyledComponentsRegistry from "@/lib/AntdRegistry";
 import { useSearchParams } from "next/navigation";
 
@@ -44,7 +43,42 @@ const DarkModeProvider: React.FC<PropsWithChildren<{}>> = ({ children }) => {
 
 }
 
+const useReportHeightToScow = () => {
+
+  useEffect(()=>{
+    // postIframeMessage();
+    const sendMessage = (height: number) => {
+      window.parent?.postMessage({
+        type: "scow.extensionPageHeightChanged",  // 发送信息的类型，不允许更改
+        payload: {
+          height: height
+        }
+      }, '*')
+    }
+
+    const observer = new ResizeObserver((entries) => {
+
+      const e = entries[0];
+      sendMessage(e.contentRect.height);
+    });
+
+
+    const htmlElement = document.querySelector("html")!;
+
+    sendMessage(htmlElement.getBoundingClientRect().height);
+
+    observer.observe(htmlElement);
+
+    return () => {
+      observer.disconnect();
+    }
+
+  }, []);
+}
+
 const Page: React.FC<React.PropsWithChildren> = ({ children }) => {
+
+  useReportHeightToScow();
 
   return (
     <App>
